split file by class
