+++ src/interface.cpp
@@ -1,6 +1,3 @@
-#include <QtGui>
-#include <QCameraInfo>
-#include <QMediaMetaData>
 #include <iostream>
 #include <cstring>
 #include <cmath>
@@ -9,6 +6,10 @@
 #include <string>
 #include <cstring>
 #include <ctime>
+
+#include <QtGui>
+#include <QCameraInfo>
+#include <QMediaMetaData>
 
 #include "pos_types.h"
 #include "interface.h"
@@ -21,43 +22,6 @@ static inline int distance(const int x1, const int y1, const int x2, const int y
 }
 
 Q_DECLARE_METATYPE(QCameraInfo)
-
-AspectRatioPixmapLabel::AspectRatioPixmapLabel(QWidget *parent) : QLabel(parent)
-{
-	this->setMinimumSize(1, 1);
-	setScaledContents(false);
-}
-
-void AspectRatioPixmapLabel::setPixmap(const QPixmap &p)
-{
-	pix = p;
-	QLabel::setPixmap(scaledPixmap());
-}
-
-int AspectRatioPixmapLabel::heightForWidth(int width) const
-{
-	if(this->height())
-		return pix.isNull();
-	else
-		return (static_cast<qreal>(pix.height()) * width) / pix.width();
-}
-
-QSize AspectRatioPixmapLabel::sizeHint() const
-{
-	int w = this->width();
-	return QSize(w, heightForWidth(w));
-}
-
-QPixmap AspectRatioPixmapLabel::scaledPixmap() const
-{
-	return pix.scaled(this->size(), Qt::KeepAspectRatio, Qt::SmoothTransformation);
-}
-
-void AspectRatioPixmapLabel::resizeEvent(QResizeEvent *e)
-{
-	if(!pix.isNull())
-		QLabel::setPixmap(scaledPixmap());
-}
 
 Interface::Interface(): fLogging(true), fReverse(false), fViewGoalpost(true), fPauseLog(false), fRecording(false), max_robot_num(6), log_speed(1), select_robot_num(-1), field_param(FieldParameter())
 {
